Generatable interface changed to ExerciseGenerator abstract class, so that it can contain a concrete method
Also had to introduce some new architecture for fill-in-the-blanks mode, so now there's a "prePrompt", which is the prompt before the question, and an "editTextPrompt", which is the prompt displayed in the EditText field. Rather than filling in the blanks, it's more like the user is completing the sentence which has been part-written. It also focuses the cursor to the correct part of the pre-written text (e.g. at the beginning, at the end, or somewhere defined in the middle). I need to resize the EditText field so it spans the width of the page and is a few lines tall (or can resize the height automatically to some set maximum), but that's for another commit.

Some general tidying of lessons code, shortening common field names

New lessons: Time, PossessionAig