Moved lessons to sub-package, created new enum GrammaticalPerson
Lessons have all been moved from the package gaidhlig to the package gaidhlig.lesssons. This is just to keep it easy to distinguish between classes which define lessons and other helper classes in gaidhlig.

Also I've created an enumerated type "GrammaticalPerson", which will replace the random number generators (as these are vulnerable to the order in which the pronouns are read from the csv files)