New lesson (SayingsComparisons) + fixed vocab files problems
Vocab files problems:
(1) Professions, for example, requires a specific vocab file. So I fixed the method which creates the set of valid vocab file names, so that it sets a specific CSV for that lesson
(2) If there's only one valid vocab file, user doesn't need to see a spinner, so that's been set to invisible, while still setting the view for vocab file size (aka number of words).

Also created a new lesson: Sayings: Comparisons